test(ui): split success/error journeys per area, add region comments, tidy names
Each feature area is now two files: <area>.spec.ts (success/CRUD) and
<area>-errors.spec.ts (validation). Files and test names use the short area name
(agents, models, mcp, prompts, chat). Every spec is organized with descriptive
// region Creating/Reading/Updating/Deleting (or Navigating/Sending/Skipping)
comments. Onboarding now verifies the created agent on the agents list it lands on.
