+++ ui/playwright/tests/agents/agents.spec.ts
@@ -3,14 +3,12 @@ import { loadPage } from "../../helpers/page";
 import { selectOption, selectNamespace } from "../../helpers/select";
 import { firstModelConfig } from "../../helpers/resources";
 
-// Agents feature — a full-CRUD lifecycle journey plus a validation-failure journey
-// (two videos). The lifecycle creates a uniquely-named declarative agent, reads it
-// back on the edit page, updates its description, and deletes it — only ever
-// touching the agent it created. The model config it attaches is discovered at
-// runtime (firstModelConfig) rather than hard-coded, and an agent is only valid
-// in the model config's namespace, so the agent is created there.
-//
-// Failure covers required-field validation on both create forms.
+// Agents — full-CRUD lifecycle journey. Creates a uniquely-named declarative
+// agent, reads it back on the edit page, updates its description, and deletes it —
+// only ever touching the agent it created. The model config it attaches is
+// discovered at runtime (firstModelConfig) rather than hard-coded, and an agent is
+// only valid in the model config's namespace, so the agent is created there.
+// Error journeys live in agents-errors.spec.ts.
 
 const DESCRIPTION = "e2e declarative agent";
 const UPDATED_DESCRIPTION = "e2e declarative agent (edited)";
@@ -21,7 +19,7 @@ async function openEdit(page: import("@playwright/test").Page, ref: string) {
   await expect(page.getByRole("heading", { level: 1, name: "Edit Agent" })).toBeVisible();
 }
 
-test("agent lifecycle: create, read, update, delete", async ({ page, request }, testInfo) => {
+test("agents: create, read, update, delete", async ({ page, request }, testInfo) => {
   const { ref: modelRef, model, namespace } = await firstModelConfig(request);
   const modelOption = `${model} (${modelRef})`;
   const name = `e2e-agent-${Date.now().toString(36)}-${testInfo.retry}`;
@@ -68,29 +66,5 @@ test("agent lifecycle: create, read, update, delete", async ({ page, request }, 
     await expect(dialog).toBeVisible();
     await dialog.getByRole("button", { name: "Delete" }).click();
     await expect(page.getByText(name)).toHaveCount(0);
-  });
-});
-
-test("agent failures: validation blocks empty required fields", async ({ page }) => {
-  // region Creating — client-side validation blocks the POST
-  await test.step("declarative create blocks submit + shows validation errors", async () => {
-    await loadPage(page, "/agents/new", { heading: "New Agent" });
-
-    await page.getByRole("button", { name: "Create Agent" }).click();
-
-    // Scroll the first error in so it's on screen (in the recorded video).
-    const descError = page.getByText("Description is required");
-    await descError.scrollIntoViewIfNeeded();
-    await expect(descError).toBeVisible();
-    await expect(page.getByText("Please select a model")).toBeVisible();
-    await expect(page).toHaveURL(/\/agents\/new/);
-  });
-
-  await test.step("harness create blocks submit when required fields are empty", async () => {
-    await loadPage(page, "/agents/new-harness", { heading: "New Agent Harness" });
-
-    await page.getByRole("button", { name: "Create harness" }).click();
-
-    await expect(page).toHaveURL(/\/agents\/new-harness/);
   });
 });
